Resolved merge conflicts across 7 files, keeping session confirmation
additions (timestamp-based capture windows, pipeline integrity counters,
transmitSessionSummary) alongside the live_debug feature code.

+++ firmware/src/components/data/DataTransmitter.h
@@ -15,7 +15,7 @@ private:
     static const size_t BATCH_SIZE = 25;             // Proximity: 25 samples per batch
     static const size_t INT_BATCH_SIZE = 100;        // Interrupt: 100 events per batch
     static const size_t LIVE_DEBUG_BATCH_SIZE = 200; // Live Debug: larger batches for speed
-    static const size_t LIVE_DEBUG_BATCH_DELAY = 20; // ms between Live Debug batches
+    static const unsigned long LIVE_DEBUG_BATCH_DELAY = 20; // ms between Live Debug batches
 
     // Session Confirmation: pointer to active session summary for transmission counters
     SessionSummary *activeSummary = nullptr;
